Cleaned up UI
