+++ WDBFontOverwrite/WDBFontOverwriteApp.swift
@@ -9,9 +9,25 @@ import SwiftUI
 
 @main
 struct WDBFontOverwriteApp: App {
-  var body: some Scene {
-    WindowGroup {
-      ContentView()
+    @StateObject private var progressManager = ProgressManager.shared
+
+    var body: some Scene {
+        WindowGroup {
+            TabView {
+                PresetFontsScene()
+                    .tabItem {
+                        Label("Presets", systemImage: "list.dash")
+                    }
+                CustomFontsScene()
+                    .tabItem {
+                        Label("Custom", systemImage: "plus")
+                    }
+                FontDiscoveryScene()
+                    .tabItem {
+                        Label("Discovery", systemImage: "star")
+                    }
+            }
+            .environmentObject(progressManager)
+        }
     }
-  }
 }
